pyproject.toml
[project]
name = "bigquery-dashboard"
version = "0.1.0"
description = "BigQuery テーブル参照分析ツール"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
  "google-cloud-bigquery>=3.0.0",
  "google-cloud-datacatalog-lineage>=0.4.0",
  "google-cloud-logging>=3.11.0",
  "pydantic>=2.12.5",
  "dash>=2.18.0,<2.19.0",
  "pandas>=2.0.0",
]

[project.scripts]
main = "app:main"
batch = "batch:main"

[dependency-groups]
dev = [
  "pyright>=1.1.407",
  "pytest>=9.0.1",
  "ruff>=0.14.6",
  "taskipy>=1.14.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dashboard", "src/shared", "src/batch", "app.py", "batch.py"]

[tool.pytest.ini_options]
addopts = [
  "-v",
  "-ra",
]
filterwarnings = [
  "ignore::DeprecationWarning",
]
python_classes = ["Test*"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.taskipy.tasks]
format = "uv run ruff format"
lint = "uv run ruff check"
test = "uv run pytest"
test-integration = "uv run pytest tests/integration -v -m integration"
test-unit = "uv run pytest tests/unit -v"
type-check = "uv run pyright"

[tool.uv]
package = true
